+++ app/triggers/providers/smtp/Smtp.js
@@ -56,14 +56,15 @@ class Smtp extends Trigger {
         await this.transporter.sendMail({
             from: this.configuration.from,
             to: this.configuration.to,
-            subject: `[WUD] New version found for image ${image.image} => ${image.result.newVersion}`,
+            subject: `[WUD] New version found for image ${image.image}`,
             text: JSON.stringify(image),
             html: `
                 <p><strong>Registry:</strong>&nbsp;${image.registry}</p>
                 <p><strong>RegistryUrl:</strong>&nbsp;${image.registryUrl}</p>
                 <p><strong>Image:</strong>&nbsp;${image.image}</p>
-                <p><strong>Current version:</strong> ${image.version}</p>
-                <p><strong>New version:</strong>&nbsp;${image.result.newVersion}</p>
+                <p><strong>Current tag:</strong> ${image.tag}</p>
+                <p><strong>Current digest:</strong> ${image.digest}</p>
+                <p><strong>New digest:</strong>&nbsp;${image.result.digest}</p>
             `,
         });
     }